webpack splits up files
