+++ .babelrc
@@ -2,20 +2,24 @@
 //     "presets": ["@babel/preset-env", "@babel/preset-react"]
 // }
 {
-    "env": {
-        "development": {
-            "presets": [
-                "@babel/preset-env",
-                "@babel/preset-react"
-            ],
-        },
-        "production": {
-            "presets": [
-                "@babel/preset-env",
-                "@babel/preset-react"
-            ]
-        }
+  "env": {
+    "development": {
+      "presets": ["@babel/preset-env", "@babel/preset-react"],
+      "plugins": [
+        "@babel/plugin-syntax-dynamic-import",
+        [
+          "import-inspector",
+          {
+            "serverSideRequirePath": true
+          }
+        ]
+      ]
+    },
+    "production": {
+      "presets": ["@babel/preset-env", "@babel/preset-react"],
+      "plugins": ["@babel/plugin-syntax-dynamic-import"]
     }
+  }
 }
 // module.exports = function () {
 //     const presets = ["@babel/preset-env", "@babel/preset-react"];
