+++ package.json
@@ -30,18 +30,28 @@
     "@babel/core": "^7.2.2",
     "@babel/node": "^7.2.2",
     "@babel/parser": "^7.2.3",
+    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/polyfill": "^7.2.5",
     "@babel/preset-env": "^7.2.3",
     "@babel/preset-react": "^7.0.0",
     "@babel/register": "^7.0.0",
     "@babel/traverse": "^7.2.3",
+    "babel-loader": "^8.0.5",
+    "babel-plugin-import-inspector": "^2.0.0",
     "babel-traverse": "^6.26.0",
     "commander": "^2.19.0",
+    "css-loader": "^2.1.0",
     "express": "^4.16.4",
+    "extract-text-webpack-plugin": "^3.0.2",
     "fs": "0.0.1-security",
+    "html-webpack-plugin": "^3.2.0",
+    "postcss-loader": "^3.0.0",
     "inquirer": "^6.2.1",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
-    "react-router-dom": "^4.3.1"
+    "react-router-dom": "^4.3.1",
+    "style-loader": "^0.23.1",
+    "webpack": "^4.28.4",
+    "webpack-cli": "^3.2.1"
   }
 }
